requirements.txt

asgiref==3.5.2
asttokens==2.0.7
backcall==0.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
debugpy==1.6.2
decorator==5.1.1
Django==4.0.8
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
entrypoints==0.4
executing==0.10.0
idna==3.4
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
mysqlclient==2.1.1
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.12.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
tornado==6.2
traitlets==5.3.0
twilio==7.14.1
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.2.5
